+++ CHANGELOG.md
@@ -16,9 +16,10 @@
 * expose kafka metrics to /_sys and collectStatTask
 
 ## Change log
-### 4.8.9 (1/12/2017 - )
+### 4.8.9 (1/12/2017 - 1/17/2017)
 * kafka: update logging to expose more info
 * mongo: check entities must not be empty in bulk operations
+* redis: support scan keys
 
 ### 4.8.8 (1/9/2017 - 1/12/2017)
 * kafka: update to 0.10.1.1